split the Z21's heartbeat from the Z21Adapter and make it only send a heartbeat if no outgoing message is sent in the specified timeframe.
