+++ java/src/jmri/jmrix/roco/z21/Z21Adapter.java
@@ -20,14 +20,6 @@ public class Z21Adapter extends jmri.jmrix.AbstractNetworkPortController {
     protected static int COMMUNICATION_UDP_PORT = java.lang.Integer.parseInt(rb.getString("z21UDPPort1"));
     protected static String DEFAULT_IP_ADDRESS = rb.getString("defaultZ21IPAddress");
 
-    private javax.swing.Timer keepAliveTimer; // Timer used to periodically
-    // send a message to both
-    // ports to keep the ports
-    // open
-    private static final int keepAliveTimeoutValue = 30000; // Interval
-    // to send a message
-    // Must be < 60s.
-
     private DatagramSocket socket = null;
 
     public Z21Adapter() {
@@ -91,8 +83,6 @@ public class Z21Adapter extends jmri.jmrix.AbstractNetworkPortController {
                     m_HostName, ConnectionStatus.CONNECTION_UP);
         }
 
-        keepAliveTimer();
-
     }
 
     /*
@@ -125,33 +115,9 @@ public class Z21Adapter extends jmri.jmrix.AbstractNetworkPortController {
     protected void resetupConnection() {
     }
 
-    /*
-     * Set up the keepAliveTimer, and start it.
-     */
-    private void keepAliveTimer() {
-        if (keepAliveTimer == null) {
-            keepAliveTimer = new javax.swing.Timer(keepAliveTimeoutValue, new java.awt.event.ActionListener() {
-                @Override
-                public void actionPerformed(java.awt.event.ActionEvent e) {
-                    // If the timer times out, send a request for status
-                    Z21Adapter.this.getSystemConnectionMemo().getTrafficController()
-                            .sendz21Message(
-                                    jmri.jmrix.roco.z21.Z21Message.getSerialNumberRequestMessage(),
-                                    null);
-                }
-            });
-        }
-        keepAliveTimer.stop();
-        keepAliveTimer.setInitialDelay(keepAliveTimeoutValue);
-        keepAliveTimer.setRepeats(true);
-        keepAliveTimer.start();
-    }
-
     @Override
     public void dispose(){
        super.dispose();
-       keepAliveTimer.stop();
-       keepAliveTimer = null;
        socket.close();
        opened = false;
        allowConnectionRecovery = false; // disposing of the object should
