+++ java/src/jmri/jmrix/roco/z21/Z21SystemConnectionMemo.java
@@ -17,7 +17,7 @@ import org.slf4j.LoggerFactory;
  * @author	Bob Jacobsen Copyright (C) 2010 copied from NCE into PowerLine for
  * multiple connections by
  * @author	Ken Cameron Copyright (C) 2011 copied from PowerLine into z21 by
- * @author	Paul Bender Copyright (C) 2013
+ * @author	Paul Bender Copyright (C) 2013,2019
  */
 public class Z21SystemConnectionMemo extends jmri.jmrix.SystemConnectionMemo {
 
@@ -206,6 +206,9 @@ public class Z21SystemConnectionMemo extends jmri.jmrix.SystemConnectionMemo {
         // setup the MultiMeter
         getMultiMeter();
 
+        // setup the HeartBeat
+        getHeartBeat();
+
    }
 
     @Override
@@ -260,6 +263,20 @@ public class Z21SystemConnectionMemo extends jmri.jmrix.SystemConnectionMemo {
 
     private Z21MultiMeter meter = null;
 
+    /**
+     * Provide access to the Z21HeartBeat instance for this connection.
+     * <p>
+     * NOTE: HeartBeat defaults to NULL
+     */
+    public Z21HeartBeat getHeartBeat() {
+        if(heartBeat == null){
+           heartBeat = new Z21HeartBeat(this);
+        }
+        return heartBeat;
+    }
+    
+    private Z21HeartBeat heartBeat = null;
+
 
     void shutdownTunnel(){
         if (_xnettunnel!=null) {
@@ -270,6 +287,7 @@ public class Z21SystemConnectionMemo extends jmri.jmrix.SystemConnectionMemo {
 
     @Override
     public void dispose() {
+        heartBeat.dispose();
         shutdownTunnel();
         InstanceManager.deregister(this, Z21SystemConnectionMemo.class);
         super.dispose();
